Adopt GPLv2 (#28)
In order to ensure that:

1. Derivative works are strongly encouraged to be released into the public domain
2. There is space for a vibrant, sustainable ecosystem around nourish.party

We believe that we should adopt the GPLv2, as that is used by several
large, open source ecosystems such as Wordpress and Drupal. We considered GPLv3, but were hesitant as we were unable to find ecosystems as vibrant as Wordpress and Drupal using the GPLv3 license

This was merged with the unanimous consent of Jennifer Tu, Betsy Haibel, Dash Buck and Zee Spencer.