Project creator may update project title and summary
This allows a project creator to update their project and automatically
resubmit it as a project that is pending curation.

Mostly boring code, with a kind of funky thing going on of switching
which notification we present to the user based on if the project was
re-submitted or simply updated.

Resolves https://github.com/wecohere/nourish/issues/55